fix: bot comments v3
