+++ api/v2/bundles/index.js
@@ -92,148 +92,10 @@ module.exports = async function handler(req, res) {
     });
   }
 
-  // Batch version of sanitizeBundle for listings — 2 parallel Redis rounds instead of 4N sequential.
-  const sanitizeBundles = async rawItems => {
-    const processed = rawItems.map(({ bundle, packageName }) => {
-      const v = bundle.min_runtime_version;
-      const min_runtime_version =
-        v != null && String(v).trim() ? String(v).trim() : '0.1.0';
-      const meta = bundle.metadata ? { ...bundle.metadata } : {};
-      const ownerEmail = (meta._ownerEmail || '').toLowerCase();
-      const hadAdminVerified = !!meta._adminVerified;
-      delete meta._ownerEmail;
-      delete meta._adminVerified;
-      return {
-        bundle,
-        packageName,
-        meta,
-        ownerEmail,
-        hadAdminVerified,
-        min_runtime_version,
-      };
-    });
-
-    const uniquePackages = [
-      ...new Set(
-        processed.map(p => p.packageName || p.bundle.package).filter(Boolean)
-      ),
-    ];
-    const uniqueEmails = [
-      ...new Set(
-        processed
-          .map(p => p.ownerEmail)
-          .filter(e => e && !e.endsWith('@calimero.network'))
-      ),
-    ];
-    const [pkgVerifiedVals, userIdVals] = await Promise.all([
-      Promise.all(
-        uniquePackages.map(p => kv.get(`admin_verified:package:${p}`))
-      ),
-      Promise.all(uniqueEmails.map(e => kv.get(`email2user:${e}`))),
-    ]);
-    const pkgVerifiedMap = Object.fromEntries(
-      uniquePackages.map((p, i) => [p, pkgVerifiedVals[i] === '1'])
-    );
-    const emailToUserId = Object.fromEntries(
-      uniqueEmails.map((e, i) => [e, userIdVals[i]])
-    );
-
-    const uniqueUserIds = [
-      ...new Set(Object.values(emailToUserId).filter(Boolean)),
-    ];
-    const [userVals, userAdminVerifiedVals] = uniqueUserIds.length
-      ? await Promise.all([
-          Promise.all(uniqueUserIds.map(id => kv.get(`user:${id}`))),
-          Promise.all(
-            uniqueUserIds.map(id => kv.get(`admin_verified:user:${id}`))
-          ),
-        ])
-      : [[], []];
-    const userMap = Object.fromEntries(
-      uniqueUserIds.map((id, i) => {
-        try {
-          return [id, userVals[i] ? JSON.parse(userVals[i]) : null];
-        } catch {
-          return [id, null];
-        }
-      })
-    );
-    const userAdminVerifiedMap = Object.fromEntries(
-      uniqueUserIds.map((id, i) => [id, userAdminVerifiedVals[i] === '1'])
-    );
-
-    return processed.map(
-      ({
-        bundle,
-        packageName,
-        meta,
-        ownerEmail,
-        hadAdminVerified,
-        min_runtime_version,
-      }) => {
-        let verified = hadAdminVerified;
-        const pkg = packageName || bundle.package;
-        if (!verified && pkgVerifiedMap[pkg]) verified = true;
-        if (!verified && ownerEmail.endsWith('@calimero.network'))
-          verified = true;
-        if (!verified && ownerEmail) {
-          const userId = emailToUserId[ownerEmail];
-          if (userId) {
-            if (userMap[userId]?.verified) verified = true;
-            if (!verified && userAdminVerifiedMap[userId]) verified = true;
-          }
-        }
-        return { ...bundle, metadata: meta, min_runtime_version, verified };
-      }
-    );
-  };
-
-  // Normalize min_runtime_version and compute `verified` server-side.
-  // Strips _ownerEmail and _adminVerified from metadata so emails are never exposed.
-  const sanitizeBundle = async (bundle, packageName) => {
-    if (!bundle || typeof bundle !== 'object') return bundle;
-    const v = bundle.min_runtime_version;
-    const min_runtime_version =
-      v != null && String(v).trim() ? String(v).trim() : '0.1.0';
-
-    const meta = bundle.metadata ? { ...bundle.metadata } : {};
-    const ownerEmail = (meta._ownerEmail || '').toLowerCase();
-    const hadAdminVerified = !!meta._adminVerified;
-    delete meta._ownerEmail;
-    delete meta._adminVerified;
-
-    let verified = hadAdminVerified;
-
-    if (!verified) {
-      const pkgKey = await kv.get(
-        `admin_verified:package:${packageName || bundle.package}`
-      );
-      if (pkgKey === '1') verified = true;
-    }
-    if (!verified && ownerEmail.endsWith('@calimero.network')) {
-      verified = true;
-    }
-    if (!verified && ownerEmail) {
-      const userId = await kv.get(`email2user:${ownerEmail}`);
-      if (userId) {
-        const userRaw = await kv.get(`user:${userId}`);
-        if (userRaw) {
-          try {
-            const user = JSON.parse(userRaw);
-            if (user.verified) verified = true;
-          } catch {
-            /* skip */
-          }
-        }
-        if (!verified) {
-          const adminVerified = await kv.get(`admin_verified:user:${userId}`);
-          if (adminVerified === '1') verified = true;
-        }
-      }
-    }
-
-    return { ...bundle, metadata: meta, min_runtime_version, verified };
-  };
+  const {
+    createBundleSanitizers,
+  } = require('../../../packages/backend/src/lib/bundle-sanitize');
+  const { sanitizeBundle, sanitizeBundles } = createBundleSanitizers(kv);
 
   try {
     const semver = require('semver');
