+++ packages/backend/src/lib/kv-client.js
@@ -134,6 +134,19 @@ if (isProduction && process.env.REDIS_URL) {
       await this._ensureConnected();
       return await redisClient.keys(pattern);
     },
+
+    /** Prefer over KEYS in production — non-blocking SCAN iterator. */
+    async scanKeys(pattern) {
+      await this._ensureConnected();
+      const result = [];
+      for await (const key of redisClient.scanIterator({
+        MATCH: pattern,
+        COUNT: 500,
+      })) {
+        result.push(key);
+      }
+      return result;
+    },
   };
 } else {
   // Mock Redis for local development using in-memory storage
@@ -254,6 +267,10 @@ if (isProduction && process.env.REDIS_URL) {
         `^${pattern.replace(/[.+^${}()|[\]\\]/g, '\\$&').replace(/\*/g, '.*')}$`
       );
       return [...mockStore.keys()].filter(k => regex.test(k));
+    },
+
+    async scanKeys(pattern) {
+      return this.keys(pattern);
     },
 
     // Utility for testing
